requirements.txt

asyncssh==2.15.0
pyyaml==6.0
aiofiles==24.1.0
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
aiohttp>=3.9.1
